+++ spring-stream/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.2
+
+- **added**: reexport `sea_streamer`
+- **added**: use `spring::submit_config_schema`
+
 ## 0.4.1
 
 - **added**: serde derive